chore!: move tracing tables to the schema package
